[renderbox] Rename metal_31001 to renderbox.py, finish removing from GCN
